requirements.txt
accelerate==0.23.0
aiohttp==3.8.6
aiosignal==1.3.1
alabaster==0.7.13
anyio==3.7.1
arrow==1.3.0
arxiv==1.4.8
async-timeout==4.0.3
attrs==23.1.0
Babel==2.13.0
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
bibtexparser==1.4.1
cchardet==2.1.7
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.3.0
Chroma==0.2.0
chroma-hnswlib==0.7.3
chromadb==0.4.14
click==8.1.7
coloredlogs==15.0.1
datasets==2.14.5
Deprecated==1.2.14
dill==0.3.7
dnspython==2.4.2
docutils==0.18.1
exceptiongroup==1.1.3
fake-useragent==1.3.0
fastapi==0.103.2
feedparser==6.0.10
filelock==3.12.4
flatbuffers==23.5.26
free-proxy==1.1.1
frozenlist==1.4.0
fsspec==2023.6.0
geographiclib==2.0
geopy==2.4.0
googlesearch-python==1.2.3
gradio_client==0.6.0
grpcio==1.59.0
h11==0.14.0
httpcore==0.18.0
httptools==0.6.0
httpx==0.25.0
huggingface-hub==0.17.3
humanfriendly==10.0
idna==3.4
imagesize==1.4.1
importlib-resources==6.1.0
iniconfig==2.0.0
jaraco.context==4.3.0
Jinja2==3.1.2
loguru==0.7.2
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
monotonic==1.6
more-itertools==10.1.0
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
networkx==3.1
numexpr==2.8.7
numpy==1.26.0
onnxruntime==1.16.1
openai==0.28.1
optimum==1.13.2
outcome==1.2.0
overrides==7.4.0
packaging==23.2
pandas==2.1.1
peft==0.5.0
pexpect==4.8.0
pinecone-client==2.2.4
pluggy==1.3.0
posthog==3.0.2
prompt-toolkit==3.0.39
protobuf==4.24.4
psutil==5.9.5
ptyprocess==0.7.0
pulsar-client==3.3.0
pyarrow==13.0.0
pydantic==1.9.0
Pygments==2.16.1
pyparsing==3.1.1
PyPika==0.48.9
PySocks==1.7.1
pytest==7.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
rich==13.6.0
safetensors==0.4.0
scholarly==1.7.11
selenium==4.14.0
sentencepiece==0.1.99
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.5
Sphinx==7.2.6
sphinx-rtd-theme==1.3.0
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
starlette==0.27.0
sympy==1.12
tenacity==8.2.3
tiktoken==0.5.1
tokenizers==0.14.1
tomli==2.0.1
torch==2.1.0
tqdm==4.66.1
transformers==4.34.0
trio==0.22.2
trio-websocket==0.11.1
typer==0.9.0
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
wcwidth==0.2.8
websockets==11.0.3
wolframalpha==5.0.0
wrapt==1.15.0
wsproto==1.2.0
xmltodict==0.13.0
xxhash==3.4.1
yarl==1.9.2
pytesseract==0.3.10
Pillow==8.3.2
requests==2.26.0

